+++ chat-react2/src/App.tsx
@@ -2,8 +2,8 @@ import { StrictMode } from "react";
 import { ChatProvider } from "@connectycube/use-chat";
 import { BrowserRouter } from "react-router-dom";
 import { Navigate, Route, Routes } from "react-router";
-import Login from "./Components/Login/Login";
-import SignUp from "./Components/SignUp/SignUp";
+import Login from "./Components/Login";
+import SignUp from "./Components/SignUp";
 import Home from "./Components/Home/Home";
 import ConnectyCube from "connectycube";
 import { appConfig, credentials } from "./config";
