--- chat-react2/src/components/Login/Login.scss
@@ -1,69 +1,0 @@
-$offwhite: darken(white, 10%);
-$darkblue: rgb(50, 50, 60);
-
-.login__container {
-  background: whitesmoke;
-  border: 1px solid black;
-  border-radius: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 40px;
-  & .img__container {
-    max-width: 200px;
-
-    & .logo__img {
-      width: 100%;
-    }
-  }
-  & .login__form {
-    display: flex;
-    flex-direction: column;
-    & .error {
-        color: red;
-      }
-    & input {
-      padding: 12px 20px;
-      border: 1px solid #eee;
-      border-radius: 25px;
-      background-color: #eee;
-      color: black;
-      font-size: 17px;
-      font-weight: 200;
-      outline: none;
-    }
-    & button {
-      border-radius: 25px;
-      border: 1px solid #282c34;
-      padding: 12px 30px;
-      color: darken($darkblue, 10%);
-      border: none;
-      background: transparent;
-      border: 1px solid black;
-      overflow: hidden;
-      box-shadow: none;
-      transition: 0.2s ease-in-out;
-      font-weight: 600;
-      margin-top: 15px;
-      margin-bottom: 20px;
-      cursor: pointer;
-      &:hover {
-        background-color: lighten($darkblue, 10%);
-        color: white;
-        border: 1px solid whitesmoke;
-      }
-    }
-    & > * {
-      &:not(:last-child) {
-        margin-bottom: 10px;
-      }
-    }
-  }
-  & .signup__block {
-    & p {
-      text-align: left;
-      font-size: 14px;
-      padding: 5px 0px;
-    }
-  }
-}
